JMathLib.jar: avoid using ant
The old version of the Fakefile tried to convince ant to compile the
classes, and wanted to combine them into a .jar file.  However, the Fiji
Build system needs to expand the wildcards, but the necessary files do
not exist the beginning, but only after running ant.

So use the Fiji Build system to compile the classes, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>