Add getComponentByPath function to testlib.
This function takes a path of the form

  PathElem>PathElem>PathElem

where PathElem again is some type-specifier of the form

  ComponentType[idx]{label}

ComponentType is the class-name of a AWT-Comonent (e.g.
Java::JavaAwt::Label or Java::JavaAwt::Button), idx is a numerical index
(starting at 0) and label is a label that can be derived from calling
getLabel on the specific component.

If idx is ommited, it is taken as 0 implicitly (thus returning the first
component of type ComponentType).  If label is ommited, testing for a
label is skipped.

Note thas this function is work-in-progress, and with a great change is
going to undergo reworking or to being replaced by a pure Java function.