Plugin-Recorder: Created logic for generic recorder
Implemented a (dummy) set of classes dedicated to making the recorder as
generic as possible.

Now a caught command will be matched by a set of rules, ordered by
priority. Rules are read from xml files, and they contain the logic to
match commands (in the form of regex) and to generate language strings
from it.

One very dummy rule catching everything and saying that it can't do
anything with it is implemented as demo.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>