Bleb-dynamics: methods to sample the shape in XY coords
Also moved it to the fiji.plugin package. Eventually in the future we
will have to move it to another sub-package.
So far the interfaces to draw and sample a shape on a list of XY pixels
is done. Now we need an interactive tool to allow the user to manipulate
it.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>