Constrained-shapes: snapping circle crude version
The snapping stuff works when the user release the mouse.
The circle is then moved and defromed in a neighborhood, and try to
accomodate to minimize the summed intensity over its perimeter.

An image testsnap.jpg is joined for testing purpose.

As now, I have troubles having the image refresh done in a correct
manner. Next steps is to move the fitter to another thread anyway.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>