Constrained-shapes: new way of drawing rois for tcs
If we want to play with ROIs, we have to extend ShapeRoi.
However, we have ot the liberty to play with its fields, since they are
private.
This caused conflicts when calling other plugin (such as Create Mask) on
our extending ROI plugin, for its internal shape deviated from the one
that was given by calling the super constructor of ShapeRoi at creation.
To circumvent this problem, I did like Johannes suggested for the Arrow
Roi, that is:

In interactive mode, we generate a new ROI every time the user modifies
it instead of modifying one ROI.
This is a bit unsatisfying aesthetically.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>