AbstractTool: check more often whether the tool should be uninstalled
Unfortunately, ImageJ's API does not tell us when the previously installed
tool was uninstalled, so we have to check in a little expensive way
whether the tool should be uninstalled.

To ease the load, we check only at certain strategic points in the
execution, such as when the mouse enters an image canvas.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>