Change the implementation of ij3d.Pipe.makeTube
The previous implementation of ij3d.Pipe.makeTube failed for
sequences of points where any two adjacent points differed only
in z.  It was difficult to change the implementation to fix this
in such a way that I was convinced it was correct.  This new
implementation works rather differently - it tries to match up
the edge points of each disc such that if the two adjacent disks
were overlaid, the ends of the radius running along their line
of intersection would be matched up.

This also slightly changes how the tangent to the line at each
point is calculated.  Previously, the tangent at point i was
considered to be the vector from i to i+1, whereas now (when a
point on either side of i exists) it will be the vector
from i-1 to i+1