Trainable Segmentation: down to 7.0 seconds/tree (from 22 originally)
When running more than one tree, each tree costs between 8.3 and 10.3
seconds, which is far less than the initial 3x cost.
There is still some collisions that must be eliminated.

The InformationGainFunction.java is pending. It has so much in common with
GiniFunction that an abstract class could do most of the work.

To compile, disable the InformationGainFunction.java file or edit it.