3D_Viewer - IJ2: Work in progress: try to write a first 3D plugin.
And let it automatically be loaded and incorporated into the menu bar,
using ImageJ2's plugin detection mechanism, which is based on sezpoz.

Unfortunately, it does not work so far. It does not even compile:

src-plugins/3D_Viewer/ij3d/plugin/ExampleV3DPlugin.java:9: not
assignable to imagej.plugin.ImageJPlugin
@Plugin(
^