Fix problem with discovery of 3D Viewer plugins.
The issue was that 3D Viewer wasn't discovering any plugins, because
they were all stored in JAR files in the plugins folder, and not present
on the system classpath. The only way for SezPoz to find such plugins is
to use ImageJ's class loader instead. So I added the ability for the
ImageJ2 PluginIndex to accept an alternative class loader when accessing
the plugin index. Now 3D Viewer successfully discovers plugins and
populates the menu bar with them.

In the future, we may need to tweak the way the singleton PluginIndex
works. As things stand, if external code creates the singleton using the
system class loader, then the 3D Viewer's later invocation of
PluginIndex.getIndex(IJ.getClassLoader()) will not discover any plugins,
since a prior index will already exist. But the current approach should
be good enough for additional plugin development, for the time being.

One remaining quirk is that there are now two Help menus, since one is
manually created while the other is built by the ImageJ2 JMenuBarCreator
logic. It should be straightforward to address it in the code, though.