add support for RGB images in NRRD reader
RGB pixels are considered as an additional dimension in the NRRD file
format, but the RGB information, stored in the "kinds" tag, can be
stored after other tags where we already have to know whether the
image is RGB or not. This force to found if the image is an RGB image before
actually parsing the header. That way, the additional dimension corresponding
to the RGB components can be removed.