Add a class to multiplex System.out/err thread-specifically
When some compiled class uses System.out or System.err to output
information, it might be cool to redirect that into, say, the output
area of the Script Editor. In most cases, this redirection is meant to
be limited to the current thread (resp. all threads spawned from it).

This class allows thread-local multiplexing of PrintStreams, and for
convenience it also includes singletons for System.out and System.err,
falling back to the previously set objects.

Since we do not want this code to get confused by Help>Refresh Menus,
the class lives in Fiji.jar (which is loaded via the system class
loader and thusly cannot be loaded twice).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>