Add a Service Provider example
The SPI (Service Provider Infrastructure) added with Java 6 allows you to
discover so-called "services", i.e. classes that instantiate certain
interfaces, without loading all classes first.

The trick is to list the classes in META-INF/services/<interfacename>
which implement the given interface.

See the example.Test class for an example to enumerate the implementing
classes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>