NucleiTracker_: implement Bhavna nuclei masks (the crown)
This is step 2 in Bhavna's Crown Wearing Nuclei Tracker.
It uses processing in the 2D planes, where the nuclei are well defined,
to compute a mask that will stress the separation between nuclei.
Unfortunately, I have found that the MATLAB implementation is faster,
particularly for the anisotropic diffusion. I will investigate that.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>