CWNT_: plugin can launch segmentation in batch
However, there is a severe NPE in the labeling of the splitted image,
when the user tries to iterate over the resulting labels.
I do not understand this NPE, but suspect a problem with holes in
labeling. Probably, re-generating the labels from scratch rather than
copying good ones and changing bad ones would be a good idea.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>