MultiViewTracker_: first version of a working opener
Virtual stack are opened, one imp per view, and transforms are
retrieved. After removing the thumbnail grabber from the annotation tool
(which caused the whole dataset to be loaded into memory in its current
version), we can annotate 12.3GB of data using 512MB.
On this demo, there still seems to be a problem with one of the
transform, but this is apparently not caused by the soft anyhow.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>